1.3-1.5
So the goal is to display the name and the parts and now things are in a array. so what changed is I still have 4 components header,parts,content,total. the last 3 have changed. in part it takes name and exercise which are passed into contnent. this is where things changed so now i have to pass the index of the part and excerise or name to get it to displaty. same thing in my total where i had to pass the prop.parts[index[.exercise to acces the number to add to the others. in my app i needed to now pass parts={parts} and exercise={excerise}